+++ dunwu-project/dunwu-data/src/main/java/io/github/dunwu/data/entity/BaseEntity.java
@@ -2,7 +2,6 @@ package io.github.dunwu.data.entity;
 
 import com.baomidou.mybatisplus.annotation.IdType;
 import com.baomidou.mybatisplus.annotation.TableId;
-import com.baomidou.mybatisplus.annotation.Version;
 import io.swagger.annotations.ApiModelProperty;
 import lombok.Data;
 
@@ -21,6 +20,6 @@ public abstract class BaseEntity implements Serializable {
 
 	@TableId(value = "id", type = IdType.AUTO)
 	@ApiModelProperty(value = "ID", example = "0")
-	protected Long id;
+	protected Integer id;
 
 }
